go/worker/keymanager/nodetracker: Refactor and move node tracker
The key manager node tracker is moved into the key manager client wrapper
and away from the P2P layer, as its main responsibility is to track nodes
rather than peers. As a result, the wrapper now has a mapping between
node IDs and their P2P identifiers at its disposal and can forward
the ID of the node to which the RPC request was sent back to the enclave.