go/worker/keymanager: Optimize enclave initialization
Enclave initialization was moved into its own goroutine to avoid blocking
the main loop of the key manager worker. Once initialization is completed,
the resulting state of the enclave is compared to the latest key manager
status. If the latter has changed, initialization is performed again.

This will be useful when we deploy master secret rotation since new secrets
may be generated while old secrets are being replicated which can result
in an outdated state once initialization finishes.