runtime/src/consensus/tendermint/verifier: Modify trust root storage key
The Tendermint verifier now stores the trusted root under a key derived
from the enclave's identity and the runtime version. Relying solely
on the identity is not sufficient, as on non-SGX platforms, the identity
is set to the default value, causing new runtimes in the upgrade tests
to use trusted roots from the previous versions, which may not even
be compatible.