runtime: Do not clamp generated X25519 keys
Even though this changes the byte-for-byte representation of returned
keys, this should have no effect on clients using them for X25519 which
will clamp internally. The only other use case is the VRF key derivation
for the RNG in the runtime SDK, but that does only uses ephemeral keys
from the latest epoch.