keymanager: Fix public key decoding for legacy key manager clients
Legacy clients failed to decode signed public keys retrieved via a secure
channel due to the uncleared expiration epoch. The problem arose because
session info isn't a reliable indicator of whether an RPC call was made
through a secure or an insecure channel. Consequently, tests with
unrestricted enclave access to the key manager, when the policy signer
set was empty or unsafe policy skip was active, failed.