Support detached ROFL components
Previously each bundle that contained one or more ROFL components also
needed to contain the exact version of the RONL component it was
attaching to.

This is somewhat awkward to use when we assume a more decentralized
development and deployment of ROFL applications. This commit adds support
for detached ROFL components where the bundle only contains the ROFL and
oasis-node then automatically gets the appropriate RONL component from
another bundle.